PokezTests: GameTest library code
This file has 3 tests the generate registrations for a RED, GREEN, or Blue Blocks. When run in the game they appear in a sucker 3x3x3 structure along with the Settings and Structure configuration blocks. Test are run as: 
/test run poketests.pokeredblocktest
/test run poketests.pokegreenblocktest
/test run poketests.pokeblueblocktest
/test runnel
The test will create a sunken 3x3x3 structure along with the poke block in the middle as well as the Structure and Config console blocks